Associate events with creator(user)

+++ app/controllers/events_controller.rb
@@ -1,10 +1,13 @@
 class EventsController < ApplicationController
+  before_action :authenticate_user!, except: [ :index, :show ]
+
   def index
-    @events = Event.all
+    @events = Event.all.includes(:creator)
   end
 
   def show
     @event = Event.find(params[:id])
+    @is_creator = @event.creator == current_user
   end
 
   def new
@@ -12,7 +15,7 @@ class EventsController < ApplicationController
   end
 
   def create
-    @event = Event.new(event_params)
+    @event = current_user.created_events.new(event_params)
 
     if @event.save
       redirect_to root_path, notice: "Success! New event created."
@@ -23,10 +26,18 @@ class EventsController < ApplicationController
 
   def edit
     @event = Event.find(params[:id])
+
+    if @event.creator != current_user
+      redirect_to root_path, alert: "You are not allowed to edit this event."
+    end
   end
 
   def update
     @event = Event.find(params[:id])
+
+    if @event.creator != current_user
+      redirect_to root_path, alert: "You are not allowed to edit this event."
+    end
 
     if @event.update(event_params)
       redirect_to @event, notice: "Event updated successfully!"
@@ -38,6 +49,10 @@ class EventsController < ApplicationController
 
   def destroy
     @event = Event.find(params[:id])
+
+    if @event.creator != current_user
+      redirect_to root_path, alert: "You are not allowed to delete this event."
+    end
 
     if @event.destroy
       redirect_to root_path, notice: "Event delete."
